Improve check of whether user is logged in or not
Previously we just checked if the uid was what we expected it to be.
This worked but was hacky and won't work with multiple users. This
commit improves the check by running a sql statement that gets the uid
and the is_admin value. Then the caller of loggedIn() can check the
permission directly.