Implement my own CSRF prevention rather than using library
I was previously using the github.com/gorilla/csrf library to prevent
CSRF attacks. This library is fine and well designed, but is
overengineered and makes something that should be simple require more
work than necessary.

Specifically there is no reason to make javascript API calls set a
CSRF prevention cookie if you set a header instead. Custom headers are
not sent cross-origin by the browser (unless specifically exempted), so
checking for the presence of a specified custom header in the CSRF
prevention code is as good as setting a token and requires less work on
both the client and server side.

This even works with cross origin requests because client code cannot
set the Origin header, so we can check that the Origin header and the
custom header match what we expect.